DO NOT REVIEW: sort: check TestAdversary
Verify that TestAdversary actually constructs an array which triggers
quadratic behaviour in Quicksort.

The adversary modifies the data on the fly in response to comparisons,
so it's not clear that the same quadratic behaviour can be triggered
when sorting a static array of ints.

This change modifies TestAdversary to construct the adversary data and
then sort it normally while counting comparisons. Heapsort has been
disabled in sort.go, so the number of comparisons should be quadratic
and the test should fail:

--- FAIL: TestAdversary (0.13s)
	sort_test.go:543: size=10000, maxcmp=560000, ncmp=10027157

Change-Id: I38188613901d5863c15648ba9cc66917c46c410f