Update patch set 2
Patch Set 2:

Should we teach ssa amd64 backend to replace (SHLQ $1, AX then SHRQ $1, AX) by BTR $63, AX? 

As far as i can read the tables of modern intel cpus and doing a quick benchmark this has 1 cycle lower latency while having the same throughput. On modern intel cpus it seems to use the same ports as shifts.

AMD Ryzen seems to have a latency of 2 and a throughput of 1 so i am not sure if this will practically cause a regression in throughput as i do not have a RYZEN to test it with.

However overall i think it could be a net win. 
If that sounds reasonable and i have not overlooked a possible performance regression i would be happy to send a CL for it.

Patch-set: 2
CC: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>