archive/tar: require opt-in to PAX format for sub-second resolution
Nearly every FileInfoHeader obtained from the FS will have timestamps
with sub-second resolution. This problematic as it now forces the PAX
format to *always* be used. This comes with two problems:
* PAX is still not as widely supported compared to USTAR
* The PAX headers will occupy at minimum 1KiB for every entry

The old behavior of tar had no support for sub-second resolution,
so it never ran into these issues. Instead, it would just truncate off
the sub-second information.

In this CL, we preserve the behavior that USTAR and GNU can still
encode timestamps with sub-second resolution, but users that care for it
will need to opt-in into it by explicitly specifying the PAX format.

Updates #11171

Change-Id: Icc5274d4245922924498fd79b8d3ae94d5717271