Update patch set 2
Patch Set 2:

I take it back - the logic is OK. I was mixing up different things in my head. boring.Enabled is only true when we are using BoringCrypto, so not on non-linux/amd64 systems. So this is only forcing the result to true arguably incorrectly on linux/amd64 systems without AES hardware support where we're nonetheless running BoringCrypto. But even there we want to prioritize AESGCM. I will send a CL as promised but only with a comment.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>