Update patch set 1
Patch Set 1:

Understanding that this is off the table atm (no worries), I thought it also worth highlighting a different concern: the vendor.json file seems to indicate that three packages from golang.org/x/arch are used, from three different revisions. dep strictly disallows that pattern, requiring that exactly one version of a project/repo be used for packages within it. (This is a complexity-managing measure, intended to simplify both the algorithm and the resulting depgraphs). I've no idea if that causes a critical problem here, or if it's possible to pick one of the three revisions that's used and the other two package will be fine.

Patch-set: 1