Update patch set 2
Patch Set 2:

> I'm pretty sure it could work; see issue 20934. :)
> 
> We could also add a sys.GOARCH check to the useAEShash check, so that on non-AMD64 systems the compiler can eliminate the branch.

I was suggesting to go a step further :)
I thought of added replaying t.something.something.alg(...) with memhash(...) without any ifs in map code. There we need no guard. Additionally we change memhash and fiends to be the names for the aes versions on amd64. e.g.:

-TEXT runtime·aeshashstr(SB),NOSPLIT,$0-24
+TEXT runtime·strhash(SB),NOSPLIT,$0-24
    MOVQ  p+0(FP), AX // ptr to string struct
    MOVQ  8(AX), CX   // length of string
    MOVQ  (AX), AX // string data
    LEAQ  ret+16(FP), DX
    TESTB $1, runtime·useAEShash(SB)
    JEQ   runtime·aeshashbody(SB)
    ...
    CALL  runtime·strhash_generic(SB)
    ...
    RET


We might only need to add enough intrinsics on amd64 to inline the above code on amd64 (not aeshashbody) to avoid a call in case of aes. But have to see how hacky that is and how much benefit we get. 

> If you're going to do this, I recommend doing issue 21539 first.
Yeah, that is how i got to the thoughts above.

Patch-set: 2
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>