Update patch set 8
Patch Set 8:

(2 comments)

What I want to avoid is the capabilities of each format interspersed across every single Header field. However, I do recognize that this is information that people need to know, and it's a pain to look them up in the PAX or GNU specs.

I put a table in the description of the formats that gives a breakdown of each of the capabilities.

Patch-set: 8