Update patch set 1
Patch Set 1:

> What is the motivation here? Is this just for speeding up compilation by ignoring some nested structs?

Sorry for not writing much, was trying to get everything mailed before I disappear.

The motivation is consistency. If 4 fields is the max we want to SSA, that should be the same whether or not those fields are in embedded structs.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>