Update patch set 1
Patch Set 1:

Ok. Leaving them as is. Maybe we can look at this again if the dynamic checks have moved out and when it looks like it makes sense to have a specialized makemap version without hint. Then it might make more sense to move them to another function to not duplicate the checks.

Patch-set: 1