Update patch set 2
Patch Set 2:

> Seems useful to have a big endian try bot if we have capacity or did the existing try bots that are big endian all support AES?

We do have another s390x build VM which we were intending to use for trybots but it is currently idle. I gave Brad SSH access to it a while ago, but it's not been a priority. There aren't currently any trybots for the other big endian platforms (ppc64/mips64/mips) either (and the mips{,64} builders seems to have disappeared entirely now...?). Not sure if there are any plans with those.

I've actually been pondering whether it would be more useful to turn the spare s390x VM into a builder for the golang.org/x repos. They aren't regularly tested on non-x86 platforms, so breakages there tend to go unnoticed for longer.

Patch-set: 2