Update patch set 4
Patch Set 4:

I don't think this kind of test is useful.  We currently feel free to change the format of Go object files.  With this kind of test, every time we do that, we will have to rebuild the testdata files to the new format.  I don't see the point of doing that.

A test that explicitly recompiles an input file and makes sure that we can detect various properties might be useful.  Although just running make.bash kind of tests that too.

Patch-set: 4
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>