Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)

I'm not sure what you mean by "pattern matching".  If you are referring to writing rules to match the statements within the RotateLeft function, that is not sufficient because the purpose of the intrinsic is to avoid the costly overhead of passing arguments and returning values in storage.  This is more expensive on ppc64le than other platforms unfortunately.

Patch-set: 2
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>