Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Maybe a toolstash compare run (or note in commit that it was done) to make sure nothing changed in the std lib?

The binaries are not byte-by-byte identical, presumably because the rules are fired in different orders. I checked a few diffs of the assembly, most of them are different ordering of instructions, or different registers got picked, or operands are combined in a different way, like
<       0x0588 01416 (/Users/cherryyz/src/go/src/fmt/print.go:1103)     ASR     $63, R3, R3
<       0x058c 01420 (/Users/cherryyz/src/go/src/fmt/print.go:1103)     AND     R2<<4, R3, R2
---
>       0x0588 01416 (/Users/cherryyz/src/go/src/fmt/print.go:1103)     LSL     $4, R2, R2
>       0x058c 01420 (/Users/cherryyz/src/go/src/fmt/print.go:1103)     AND     R3->63, R2, R2

which, I think, are ok.

Patch-set: 1
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>