Update patch set 2
Patch Set 2:

(3 comments)

The new examples are good, I don't think the paragraph about index expressions is the right place to implement the fix: it's not where the bug is. That is, the new text in the CL is special-casing shift expressions used as indices, but there's a more general bug here not being fixed. I think we all agree that the intent is to allow not just x[1.0<<s] but also expressions like x[(1.0<<s)+1], and the text as written does not address the latter.

The existing index expression text here talks about "the type [the untyped constant] would assume if the shift expression were replaced by its left operand alone", and the section on untyped constants says "A constant may be given a type explicitly by a constant declaration or conversion, or implicitly when used in a variable declaration or an assignment or as an operand in an expression." Therefore if wording is missing from the spec, it's in the sections about the two expressions in question - index expressions and make expressions - to make clear that a constant appearing there is given a type explicitly by its appearance in that context.

I would suggest striking the new text in "index expressions" and instead making the changes noted in my comments below. The examples in "index expressions" are good and should stay.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>