reflect: fix pointer pass-the-end in Call with zero-sized return value
If a function with nonzero frame but zero-sized return value is
Call'd, we may write a pass-the-end pointer in preparing the
return Values. Fix by return the zero value for zero-sized
return value.

The test isn't ideal. It does seem to fail reliably without the
fix, though. Any better way to write the test?

Fixes #21717.

Change-Id: I5351cd86d898467170a888b4c3fc9392f0e7aa3b