Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > ret := reflect.ValueOf(func() (struct{}, struct{}) { return struct{}{}, struct{}{} }).Call(nil)
> > > if ret[0] == ret[1] {
> > > 	fmt.Println("Gotcha!")
> > > }
> > 
> > I don't think that works: after the fix, ret[0] and ret[1] may reasonably be byte-for-byte equal anyway (they may both have nil ptr fields).
> 
> The test may directly compare `ptr` fields of `ret`. First with `nil` and then with themselves.

That's still overspecified. The spec explicitly allows that “Pointers to distinct zero-size variables may or may not be equal”, so Zero may legitimately return the same (non-nil) ptr for each invocation on a zero-size type.

The property we want is “the ptr fields are either nil or valid pointers”, which is strictly weaker than “the ptr fields are either nil or distinct pointers”.

If the Interface method preserves invalid pointers, it might be possible to implement a more robust test along the lines you suggest by placing a non-zero-size value in between the two structs and checking that the resulting interfaces are actually equal:

ret := reflect.ValueOf(func() (struct{}, int, struct{}) {
	return struct{}{}, 42, struct{}{}
}).Call(nil)
if ret[0].Interface() != ret[2].Interface() {
	… // Fail
}

Patch-set: 1