Update patch set 1
Patch Set 1:

> Instead, can you try creating an info symbol in the compiler and emitting it into the object file fully formed? 

The problem with that is that, for some reason, the compiler doesn't know the package name, for functions it emits "".functionName and then the linker fixes it with a ugly hack. I'm not sure why that is but we would have the same problem here.

Patch-set: 1
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>