crypto/elliptic: drop s390x assembly.
This is hopefully temporary. #20215 needs to be fixed, but I cannot
update nor test any changes to this assembly so someone who maintains
s390x would need to revert this change, hopefully before the next
release, and include the needed fixes.

This is being split into a separate change to make undoing it easier.

Change-Id: I254f59c46ac4e6b5b4dcb8be5db8e8469aedf8f3