Update patch set 1
Patch Set 1:

If possible I'd prefer CL 62292 to avoid messing with the git history too much. Up to you and Russ though, this CL passes the tests on s390x.

We are definitely going to fix this before the 1.10 freeze, this is an important piece of code for us.

Patch-set: 1
CC: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>