Update patch set 2
Patch Set 2:

As for the race detector, I am pretty sure the new code is incorrect. It is introducing a new happens-before edge from a net.Conn.SetDeadline call to the return from a net.Conn.Read/Write that hits a deadline expiry. If the SetDeadline happens in an unrelated goroutine from the Read/Write, the memory model does not guarantee a happens-before relation, so the edge should not be introduced.

Patch-set: 2