runtime,reflect: arena allocation API experiment
DO NOT REVIEW
DO NOT SUBMIT
This is a proof-of-concept experiment to play with.
It is based on Austin's original CL, with a few mods.
It's not documented or fully tested.

This adds a simple, type-safe arena allocation API. This is intended
for things like deserializers of irregular structures, which often
have to construct complex and irregular object graphs, all with very
similar lifetimes.

Change-Id: Id3e2015c6748e8811240abeb2beef1fe95a413df