Update patch set 1
Patch Set 1:

> Yes, there is https://go-review.googlesource.com/c/go/+/63330
> I guess it should switch to lower -m level, if we want to use it in tests, because -m x4 generates a LOT of output. I have no opinion about pipe/scanner.

I tried x3 and it was already quite a lot. In theory we don't care that much, and we might even use other parts of the extra info. In practice, x3 and x4 could slow down the test quite a bit.

In any case, use the -m level that feels best for a human user, and we can always modify the level later if needed for this test.

Patch-set: 1
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>