Update patch set 1
Patch Set 1:

> > I don't think that's plausible, so I'm going in the other direction: eliminating it.
> 
> Perhaps, we have another option: replacing it, which is more natural to me.
>
> > It is not possible to write expressions of type `void` in C code
> 
> Yes, I remember that we talk about `incomplete types` at #19487.

That's an interesting point, but I don't think the analogy between void and incomplete types quite holds up.

In a pure Go program, the type `*T` is always a pointer to some Go type `T`, and every Go type has a well-defined zero-value. For the case of incomplete types, we don't know what the zero-value is, but it is usually still defined somewhere in the program.

On the other hand, void does not have a defined value anywhere in the program. If we admit the type *C.void, then we either have a zero-value for the C void type (disallowed by the C spec) or a translation that is no longer a pointer to any (even theoretical) Go type.

Because void has no values, void* always has the meaning “pointer to externally-specified type” — and that's exactly the meaning of Go's unsafe.Pointer type. With incomplete types in general we're kind of stuck, because there is no Go type with the same meaning (“pointer to struct of unknown size with tag T”), but with unsafe.Pointer we have a viable, idiomatic alternative.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>