Update patch set 2
Patch Set 2:

Busy week and also I didn't want to make more work for David on vacation :)

> After further thought, blocks might be right after all.  I'm thinking about the case of loop headers, where a single line contains both the (usually constant) assignment as well as the increment.

Yeah. In general, it would be nice if variable state was the same for every instruction in the same line, but I don't think it's true enough for this to be a better approach, especially since it eventually has to map back to PCs to be useful to the debugger.

> ...which is empty. Is it because the variable was optimized out entirely? Shouldn't those variables be pruned entirely? It does get pruned when I compile with master.

That was an incidental side effect, but it didn't seem obviously bad to me to leave a record that the variable existed -- it's not great UI to show nothing, right? It would certainly be better for binary size to prune them though, and the empty location list is silly. Opinions welcome.

> 2. I'm also still finding some weird holes in the loclists, 

Right. Thanks. Apparently in the last couple months I forgot that blocks could be reordered, and the traversal that goes in block ID order in pgen.go is completely busted. The functions I tested with happened to not get reordered. I'll try to get that fixed next week.

Patch-set: 2
Reviewer: Gerrit User 17090 <17090@62eb7196-b449-3ce5-99f1-c037f21e1705>