Update patch set 1
Patch Set 1:

The syso is updated because only the symbols listed in the .h file are left exported after the build. So adding the new .h declaration also requires a rebuild to get the symbol exported. (That's also why the delta is so small.)

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>