Update patch set 2
Patch Set 2:

> Because it reduces duplication. Go code already can and does refer to `unsafe.Pointer` — which has the same meaning as C's `void*` — so what's the point in having a second way to spell that type?

Because the current spec permit that implicitly, it doesn't work now though.

> On the other hand, accepting `*C.void` would imply that `C.void` also corresponds to a valid Go type, which should not be the case.

I don't understand this statement. C allows to instantiate void*, but disallow to instantiate void.
 
> Furthermore, unsafe.Pointer is part of the Go spec, not just a special cgo-ism, so preferring unsafe.Pointer over `*C.void` moves cgo code to be closer to actual Go, as opposed to a Go-flavored syntax for C.

I expect that CGO inherit some amount of C semantics which is fine to me.
Rejecting C.void is reasonable because C semantics doesn't permit that.
However, C allows to instantiate *C.void. It's not obvious to me that the reason you introduce the new rule and why it is worthy.
Additionally, the fix is not consistent for me. As far as I know, C disallows void because it is a incomplete type. The fix doesn't affect other incomplete types, it deals void specially without any reason.

Patch-set: 2