Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Because it reduces duplication. Go code already can and does refer to `unsafe.Pointer` — which has the same meaning as C's `void*` — so what's the point in having a second way to spell that type?
> > > 
> > > Because the current spec permit that implicitly, it doesn't work now though.
> > 
> > The current documentation for cgo says:
> > “The C type void* is represented by Go's unsafe.Pointer.”
> > (https://golang.org/cmd/cgo/#hdr-Go_references_to_C)
> > 
> > That does not seem to permit the C type `void*` to be represented by anything else.
> 
> It doesn't permit void* explicitly. It also doesn't reject void* explicitly.

It specifies that the C type `void*` is represented by Go's `unsafe.Pointer`. It does not specify any other representation for the type `void*`, and the type `*C.void` is currently not equal to `unsafe.Pointer`, which implies that it is not a valid representation for the C type `void*`.


> > > Additionally, the fix is not consistent for me. As far as I know, C disallows void because it is a incomplete type. The fix doesn't affect other incomplete types, it deals void specially without any reason.
> > 
> > Other incomplete types do not have existing, explicitly-documented semantics in the cgo tool.
> 
> Could you give me a pointer?

Search https://tip.golang.org/cmd/cgo/ for the string “incomplete”. It's not there: the cgo tool has explicit semantics for `void*`, but not for incomplete types in general.

Patch-set: 2
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>