cmd/internal/objabi: allow more than one directory in trimpath
Enables to use the same format in `-trimpath` as in GOPATH, and to
set `-trimpath` multiple times to strip more than one prefix.

This is especially handy for reproducible builds in case GOPATH is
actually not only a single directory, as is allowed per pathspec.
Without this you could strip all the prefixes with, for example:
`GOPATH=/theirs:/ours`

Although the guarantee is that one of the matching prefixes is picked
indeed the first longest one is, to account for settings like
`-trimpath="$GOPATH:$GOPATH/src/…/vendor"`.

Change-Id: Ida2d256c51cd0e1313144e0ca7cfdcc4587220ac