Update patch set 3
Patch Set 3:

I'm still seeing holes, argNum of fmt.(*pp).doPrintln has this loclist:

    00056b46 ffffffffffffffff 0000000000487ce0 (base address)
    00056b56 0000000000487d1b 0000000000487d58 (DW_OP_call_frame_cfa)
    00056b69 0000000000487d72 0000000000487e5e (DW_OP_call_frame_cfa)

has no address for 0x487d69 which is the if argNum > 0 line.

Not for this CL but I just noticed that for fmt.Println no entry is emitted the argument, this doesn't happen on master either, I wonder why.

Patch-set: 3